apps/backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
websockets
google-genai
python-dotenv
pydantic-settings